Consistency: canonical selector javadoc, test names, add Netty client selector test

Copilot-Session: f869d1c8-483d-4427-9f86-03401a8f0f25
