+++ instrumentation/ratpack/ratpack-1.7/library/src/main/java/io/opentelemetry/instrumentation/ratpack/v1_7/RatpackClientTelemetryBuilder.java
@@ -45,12 +45,14 @@ public final class RatpackClientTelemetryBuilder {
   /**
    * Configures which HTTP request headers are captured as span attributes.
    *
+   * <p>Header values are captured under the {@code http.request.header.<key>} attribute key. The
+   * {@code <key>} part in the attribute key is the lowercase header name.
+   *
    * <p>Selector patterns are matched case-insensitively, since HTTP header names are
    * case-insensitive. {@code ?} matches one character and {@code *} matches any number of
    * characters, including none. Excluded patterns take precedence over included patterns. A
-   * selector with no included patterns captures every header that is not excluded. No headers are
-   * captured when no selector is configured or when the selector is {@linkplain
-   * IncludeExclude#isEmpty() empty}.
+   * selector with no included patterns captures every header that is not excluded, and an
+   * {@linkplain IncludeExclude#isEmpty() empty} selector captures no headers.
    */
   @CanIgnoreReturnValue
   public RatpackClientTelemetryBuilder setRequestHeaders(IncludeExclude requestHeaders) {
@@ -61,8 +63,8 @@ public final class RatpackClientTelemetryBuilder {
   /**
    * Configures HTTP request headers to capture as span attributes.
    *
-   * <p>The header names are matched exactly. Names containing {@code *} or {@code ?} are ignored
-   * and reported, since this setting never supported wildcards.
+   * <p>The header names are matched literally, so {@code *} and {@code ?} are not treated as glob
+   * patterns.
    *
    * @param requestHeaders HTTP header names to capture.
    * @deprecated Use {@link #setRequestHeaders(IncludeExclude)} instead, which matches glob patterns
@@ -79,12 +81,14 @@ public final class RatpackClientTelemetryBuilder {
   /**
    * Configures which HTTP response headers are captured as span attributes.
    *
+   * <p>Header values are captured under the {@code http.response.header.<key>} attribute key. The
+   * {@code <key>} part in the attribute key is the lowercase header name.
+   *
    * <p>Selector patterns are matched case-insensitively, since HTTP header names are
    * case-insensitive. {@code ?} matches one character and {@code *} matches any number of
    * characters, including none. Excluded patterns take precedence over included patterns. A
-   * selector with no included patterns captures every header that is not excluded. No headers are
-   * captured when no selector is configured or when the selector is {@linkplain
-   * IncludeExclude#isEmpty() empty}.
+   * selector with no included patterns captures every header that is not excluded, and an
+   * {@linkplain IncludeExclude#isEmpty() empty} selector captures no headers.
    */
   @CanIgnoreReturnValue
   public RatpackClientTelemetryBuilder setResponseHeaders(IncludeExclude responseHeaders) {
@@ -95,8 +99,8 @@ public final class RatpackClientTelemetryBuilder {
   /**
    * Configures HTTP response headers to capture as span attributes.
    *
-   * <p>The header names are matched exactly. Names containing {@code *} or {@code ?} are ignored
-   * and reported, since this setting never supported wildcards.
+   * <p>The header names are matched literally, so {@code *} and {@code ?} are not treated as glob
+   * patterns.
    *
    * @param responseHeaders HTTP header names to capture.
    * @deprecated Use {@link #setResponseHeaders(IncludeExclude)} instead, which matches glob
